requirements.txt
# Core dependencies
PyQt6
numpy
scipy
matplotlib
sounddevice
soundfile
pydub
audioop-lts

# Configuration
pydantic-settings
python-dotenv

# API
fastapi[standard]
python-multipart
httpx

# Auth & Payments
PyJWT
cryptography
stripe

# Testing
pytest
pytest-cov

# DiagBot chat: in-process model (no Ollama required). Model auto-downloads on first use.
# Windows without C++: pip install llama-cpp-python --extra-index-url https://abetlen.github.io/llama-cpp-python/whl/cpu
llama-cpp-python

# Optional: LLM providers (uncomment as needed)
# openai
# anthropic
# tavily-python

# Dispatch flow: LangGraph multi-agent flow (diagnostics -> parts -> mechanic)
langgraph>=0.2.0
langchain-core>=0.3.0

# Repair guides (crawler + cloud DB): for scripts and when REPAIR_GUIDES_DB_URL is set
beautifulsoup4
psycopg2-binary

# PDF export for diagnosis reports
reportlab

# Web Push (optional): for mechanic job notifications
# pywebpush
